Ensure debug_fly_llm runs without (new) problems
