+++ flyllm/plotting.py
@@ -4,18 +4,18 @@ import matplotlib.pyplot as plt
 from matplotlib import cm, colors
 import torch
 
+from apf.models import criterion_wrapper
+from apf.utils import npindex
+from apf.data import get_batch_idx, split_data_by_id, select_bin_edges, get_real_agents
 from flyllm.config import (
     SENSORY_PARAMS, ARENA_RADIUS_MM,
     posenames, keypointnames, scalenames, skeleton_edges, keypointidx,
     featglobal, featrelative, kpvision_other,
     nglobal, nrelative, nkptouch, nfeatures
 )
-from apf.data import get_batch_idx, load_and_filter_data, split_data_by_id, select_bin_edges, get_real_agents
 from flyllm.features import compute_features, zscore, unzscore
 from flyllm.pose import FlyExample
-from apf.models import criterion_wrapper
-from apf.utils import npindex
-from flyllm.io import read_config
+from flyllm.io import read_config, load_and_filter_data
 
 
 def select_featidx_plot(train_dataset, ntspred_plot, ntsplot_global=None, ntsplot_relative=None):
